Put external module support into a separate file
