+++ sgrpg.js
@@ -1,6 +1,7 @@
 import { SGRPG } from "./module/config.js";
 
-import "./module/scene-config.js"; // To activate the hook
+import "./module/scene-config.js";
+import "./module/external-support.js"; // To activate hooks
 
 import ItemSg from "./module/item/entity.js"
 import ActorSg from "./module/actor/entity.js"
@@ -87,34 +88,3 @@ function getTensionDie() {
     const tensionDie = game.scenes.active.getFlag("sgrpg", "sceneTensionDie") || game.settings.get("sgrpg", "campaignTension");
     return tensionDie;
 }
-
-
-/* -------------------------------------------- */
-/*  External Module Support                     */
-/* -------------------------------------------- */
-
-// Drag Ruler integration
-Hooks.once("dragRuler.ready", (SpeedProvider) => {
-    class SGSpeedProvider extends SpeedProvider {
-        get colors() {
-            return [
-                { id: "walk", default: 0x0000FF, name: "Walking Speed" },
-                { id: "dash", default: 0x00DE00, name: "Dashing Speed" }
-            ];
-        }
-
-        getRanges(token) {
-            const walkspeed = token.actor?.data.data.speed || 0;
-            const dashspeed = walkspeed * 2;
-
-            const ranges = [
-                { range: walkspeed, color: "walk" },
-                { range: dashspeed, color: "dash" }
-            ];
-
-            return ranges;
-        }
-    }
-
-    dragRuler.registerSystem("sgrpg", SGSpeedProvider);
-});
